Updated yaml + sampling (with divergences)
It is important to update seaborn to 0.13.2 in order to get this to run without bugs. Beyond that, I moved from a t distribution to a normal distribution so that the prior predictive would look reasonable. I can change back if I want. Now the prior predictives all look ok, but the posterior predictive samples with many, many divergences.